Companion `visual_search` renamed `visual_query`
This renaming has affected the inclusion path for certain files, as well
as the override for the feature flag being disabled, and the name of the
feature flag itself.

Chromium change:
https://chromium.googlesource.com/chromium/src/+/a473a7e73893c658a317522fdd6d71ae6dd102e0

commit a473a7e73893c658a317522fdd6d71ae6dd102e0
Author: Serena Pascual <srna@google.com>
Date:   Thu Nov 30 19:22:36 2023 +0000

    Rename Visual Search references in common/ code, including namespace

    Excludes changes to protos for optimization target and VisualSearchModelMetadata (server-side changes needed first).

    Bug: b/289068088
